liquidlife_backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
psycopg[binary]>=3.1,<4.0
django-cors-headers>=4.4,<5.0
djangorestframework-simplejwt>=5.3,<6.0
google-auth>=2.38,<3.0
mammoth>=1.8,<2.0
pypdf>=5.3,<6.0
